+++ airflow/executors/base_executor.py
@@ -184,7 +184,7 @@ class BaseExecutor(LoggingMixin):
         self.queue_command(
             task_instance,
             command_list_to_run,
-            priority=task_instance.priority_weight,
+            priority=task_instance.task.priority_weight_total,
             queue=task_instance.task.queue,
         )