+++ airflow/providers/elasticsearch/CHANGELOG.rst
@@ -27,6 +27,19 @@
 Changelog
 ---------
 
+5.3.0
+.....
+
+.. note::
+  This release of provider is only available for Airflow 2.6+ as explained in the
+  `Apache Airflow providers support policy <https://github.com/apache/airflow/blob/main/PROVIDERS.rst#minimum-supported-version-of-airflow-for-community-managed-providers>`_.
+
+Misc
+~~~~
+
+* ``Bump minimum Airflow version in providers to Airflow 2.6.0 (#36017)``
+* ``Cleanup code for elasticsearch<8 (#35707)``
+
 5.2.0
 .....